Refactoring unit tests. Adding express test for route expiration. Updating modules
