+++ lib/ExpressRedisCache.js
@@ -70,7 +70,7 @@ module.exports = (function () {
      *  @type Object (preferably a client from the official Redis module)
      */
 
-    this.client = this.options.client || require('redis').createClient(this.port, this.host);
+    this.client = this.options.client || require('redis').createClient(this.port, this.host, {});
 
     /** If client can emit */
 
@@ -100,7 +100,7 @@ module.exports = (function () {
   /**  js-comment
    *
    *  @method
-   *  @description Get - 
+   *  @description Get -
    *  @return void{Object}
    *  @arg {Object} arg - About arg
    */
@@ -125,7 +125,7 @@ module.exports = (function () {
    *  @arg {Object} arg - About arg
    */
 
-  ExpressRedisCache.prototype.del = require('./ExpressRedisCache/del');;
+  ExpressRedisCache.prototype.del = require('./ExpressRedisCache/del');
 
   /**  js-comment
    *
@@ -157,7 +157,7 @@ module.exports = (function () {
 
   ExpressRedisCache.init = function (options) {
     return new ExpressRedisCache(options);
-  }
+  };
 
   return ExpressRedisCache;
