+++ package.json
@@ -7,7 +7,7 @@
     "express-redis-cache": "bin/express-redis-cache.js"
   },
   "scripts": {
-    "test": "mocha --check-leaks -R nyan test test/ExpressRedisCache"
+    "test": "mocha --check-leaks -R nyan test test"
   },
   "repository": {
     "type": "git",
@@ -38,8 +38,8 @@
   "license": "BSD",
   "dependencies": {
     "async": "~0.9.0",
-    "colors": "~0.6.2",
-    "redis": "^0.11.0"
+    "colors": "~1.0.3",
+    "redis": "^0.12.1"
   },
   "directories": {
     "test": "test"
@@ -49,12 +49,12 @@
     "type": "text/html"
   },
   "devDependencies": {
-    "body-parser": "~1.5.0",
-    "method-override": "~2.1.1",
-    "request": "~2.37.0",
-    "express": "~4.9.5",
-    "mocha": "~2.0.1",
-    "should": "~4.4.1",
-    "moment": "^2.7.0"
+    "body-parser": "~1.12.0",
+    "method-override": "~2.3.1",
+    "request": "~2.53.0",
+    "express": "~4.11.2",
+    "mocha": "~2.1.0",
+    "should": "~5.0.0",
+    "moment": "^2.9.0"
   }
 }
