+++ test/get.js
@@ -8,8 +8,6 @@
   var mocha     =   require('mocha');
   var should    =   require('should');
 
-  var config    =   require('../../package.json').config;
-
   var prefix    =   'erct:';
   var host      =   'localhost';
   var port      =   6379;
@@ -18,7 +16,7 @@
   var _body     =   'test1 test1 test1';
   var _type     =   'text/plain';
 
-  var cache     =   require('../../')({
+  var cache     =   require('../')({
     prefix: prefix,
     host: host,
     port: port
@@ -41,7 +39,7 @@
     });
 
     it ( 'should not have error', function () {
-      should(error).be.undefined;
+      should(error).be.null;
     });
 
     it ( 'should be an array', function () {
